// ui/src/common/enums.ts
export enum LoadingState {
    Overview,
    Fetching,
    ShowingResults,
    WaitingInQueue,
}

export enum Representation {
    Cartoon = "cartoon",
    Tube = "tube",
    Surface = "surface",
    BallAndStick = "ball+stick",
    Spacefill = "spacefill",
    Trace = "trace",
    Licorice = "licorice",
    All = "_",
}